requirements.txt
pandas
numpy
streamlit_extras
plotly